fix(test): certify transplanted Phase 3 slices
gamemd sources: MapClass::Resize @ 0x00565C10 and CellClass ctor @ 0x0047BBF0 for cleared fallback state; LogicClass::PerTickUpdate @ 0x0055AFB0 / TeamClass rung @ 0x0055B502..0x0055B59F for scheduler order; CellClass+0x140 anchor gating from SetBridgeDirection @ 0x0047E040/0x0047E470. Refresh stale fixtures and composition-only hash ratchets, and move SHA-256 to util so sim retail-oracle tests preserve the layer boundary.
