+++ src/app/diagnostics/tactical_capture/integrity.rs
@@ -4,7 +4,6 @@
 //! file checks, stable reads, and streaming SHA-256. It has no app or sim
 //! authority.
 
-use std::fmt::Write as _;
 use std::fs::{self, File};
 use std::io::{BufReader, Read};
 use std::path::{Path, PathBuf};
@@ -14,6 +13,8 @@ use anyhow::{Context, Result, bail, ensure};
 use serde::de::{DeserializeSeed, MapAccess, SeqAccess, Visitor};
 use serde::{Deserialize, Deserializer};
 use serde_json::{Map, Value};
+
+use crate::util::sha256::{Sha256, digest_hex};
 
 const MAX_JSON_BYTES: u64 = 16 * 1024 * 1024;
 
@@ -169,21 +170,11 @@ pub(crate) fn sha256_file(path: &Path, label: &str) -> Result<FileDigest> {
 }
 
 pub(crate) fn sha256_bytes(bytes: &[u8]) -> [u8; 32] {
-    let mut sha = Sha256::new();
-    sha.update(bytes);
-    sha.finalize()
+    crate::util::sha256::sha256_bytes(bytes)
 }
 
 pub(crate) fn sha256_hex(bytes: &[u8]) -> String {
-    digest_hex(sha256_bytes(bytes))
-}
-
-fn digest_hex(digest: [u8; 32]) -> String {
-    let mut output = String::with_capacity(64);
-    for byte in digest {
-        write!(&mut output, "{byte:02x}").expect("writing into String cannot fail");
-    }
-    output
+    crate::util::sha256::sha256_hex(bytes)
 }
 
 pub(crate) fn parse_strict_json<T>(bytes: &[u8], label: &str) -> Result<T>
@@ -307,195 +298,6 @@ impl<'de> Visitor<'de> for StrictJsonVisitor {
             values.insert(key, value.0);
         }
         Ok(StrictJson(Value::Object(values)))
-    }
-}
-
-struct Sha256 {
-    state: [u32; 8],
-    block: [u8; 64],
-    block_len: usize,
-    total_len: u64,
-}
-
-impl Sha256 {
-    fn new() -> Self {
-        Self {
-            state: [
-                0x6a09_e667,
-                0xbb67_ae85,
-                0x3c6e_f372,
-                0xa54f_f53a,
-                0x510e_527f,
-                0x9b05_688c,
-                0x1f83_d9ab,
-                0x5be0_cd19,
-            ],
-            block: [0; 64],
-            block_len: 0,
-            total_len: 0,
-        }
-    }
-
-    fn update(&mut self, mut bytes: &[u8]) {
-        self.total_len = self
-            .total_len
-            .checked_add(bytes.len() as u64)
-            .expect("SHA-256 input length overflow");
-        if self.block_len != 0 {
-            let count = (64 - self.block_len).min(bytes.len());
-            self.block[self.block_len..self.block_len + count].copy_from_slice(&bytes[..count]);
-            self.block_len += count;
-            bytes = &bytes[count..];
-            if self.block_len < 64 {
-                return;
-            }
-            let block = self.block;
-            self.compress(&block);
-            self.block_len = 0;
-        }
-        while bytes.len() >= 64 {
-            let (block, remaining) = bytes.split_at(64);
-            let mut owned = [0_u8; 64];
-            owned.copy_from_slice(block);
-            self.compress(&owned);
-            bytes = remaining;
-        }
-        self.block[..bytes.len()].copy_from_slice(bytes);
-        self.block_len = bytes.len();
-    }
-
-    fn finalize(mut self) -> [u8; 32] {
-        let bit_length = self
-            .total_len
-            .checked_mul(8)
-            .expect("SHA-256 bit length overflow");
-        self.block[self.block_len] = 0x80;
-        self.block_len += 1;
-        if self.block_len > 56 {
-            self.block[self.block_len..].fill(0);
-            let block = self.block;
-            self.compress(&block);
-            self.block = [0; 64];
-        } else {
-            self.block[self.block_len..56].fill(0);
-        }
-        self.block[56..64].copy_from_slice(&bit_length.to_be_bytes());
-        let block = self.block;
-        self.compress(&block);
-        let mut output = [0_u8; 32];
-        for (chunk, word) in output.chunks_exact_mut(4).zip(self.state) {
-            chunk.copy_from_slice(&word.to_be_bytes());
-        }
-        output
-    }
-
-    fn compress(&mut self, block: &[u8; 64]) {
-        const K: [u32; 64] = [
-            0x428a_2f98,
-            0x7137_4491,
-            0xb5c0_fbcf,
-            0xe9b5_dba5,
-            0x3956_c25b,
-            0x59f1_11f1,
-            0x923f_82a4,
-            0xab1c_5ed5,
-            0xd807_aa98,
-            0x1283_5b01,
-            0x2431_85be,
-            0x550c_7dc3,
-            0x72be_5d74,
-            0x80de_b1fe,
-            0x9bdc_06a7,
-            0xc19b_f174,
-            0xe49b_69c1,
-            0xefbe_4786,
-            0x0fc1_9dc6,
-            0x240c_a1cc,
-            0x2de9_2c6f,
-            0x4a74_84aa,
-            0x5cb0_a9dc,
-            0x76f9_88da,
-            0x983e_5152,
-            0xa831_c66d,
-            0xb003_27c8,
-            0xbf59_7fc7,
-            0xc6e0_0bf3,
-            0xd5a7_9147,
-            0x06ca_6351,
-            0x1429_2967,
-            0x27b7_0a85,
-            0x2e1b_2138,
-            0x4d2c_6dfc,
-            0x5338_0d13,
-            0x650a_7354,
-            0x766a_0abb,
-            0x81c2_c92e,
-            0x9272_2c85,
-            0xa2bf_e8a1,
-            0xa81a_664b,
-            0xc24b_8b70,
-            0xc76c_51a3,
-            0xd192_e819,
-            0xd699_0624,
-            0xf40e_3585,
-            0x106a_a070,
-            0x19a4_c116,
-            0x1e37_6c08,
-            0x2748_774c,
-            0x34b0_bcb5,
-            0x391c_0cb3,
-            0x4ed8_aa4a,
-            0x5b9c_ca4f,
-            0x682e_6ff3,
-            0x748f_82ee,
-            0x78a5_636f,
-            0x84c8_7814,
-            0x8cc7_0208,
-            0x90be_fffa,
-            0xa450_6ceb,
-            0xbef9_a3f7,
-            0xc671_78f2,
-        ];
-        let mut words = [0_u32; 64];
-        for (index, chunk) in block.chunks_exact(4).enumerate() {
-            words[index] = u32::from_be_bytes(chunk.try_into().expect("four-byte chunk"));
-        }
-        for index in 16..64 {
-            let s0 = words[index - 15].rotate_right(7)
-                ^ words[index - 15].rotate_right(18)
-                ^ (words[index - 15] >> 3);
-            let s1 = words[index - 2].rotate_right(17)
-                ^ words[index - 2].rotate_right(19)
-                ^ (words[index - 2] >> 10);
-            words[index] = words[index - 16]
-                .wrapping_add(s0)
-                .wrapping_add(words[index - 7])
-                .wrapping_add(s1);
-        }
-        let [mut a, mut b, mut c, mut d, mut e, mut f, mut g, mut h] = self.state;
-        for index in 0..64 {
-            let sum1 = e.rotate_right(6) ^ e.rotate_right(11) ^ e.rotate_right(25);
-            let choice = (e & f) ^ ((!e) & g);
-            let temp1 = h
-                .wrapping_add(sum1)
-                .wrapping_add(choice)
-                .wrapping_add(K[index])
-                .wrapping_add(words[index]);
-            let sum0 = a.rotate_right(2) ^ a.rotate_right(13) ^ a.rotate_right(22);
-            let majority = (a & b) ^ (a & c) ^ (b & c);
-            let temp2 = sum0.wrapping_add(majority);
-            h = g;
-            g = f;
-            f = e;
-            e = d.wrapping_add(temp1);
-            d = c;
-            c = b;
-            b = a;
-            a = temp1.wrapping_add(temp2);
-        }
-        for (slot, value) in self.state.iter_mut().zip([a, b, c, d, e, f, g, h]) {
-            *slot = slot.wrapping_add(value);
-        }
     }
 }
