+++ src/sim/bridge_state/tests.rs
@@ -1554,7 +1554,14 @@ fn make_bridgehead_terrain_ns() -> crate::map::resolved_terrain::ResolvedTerrain
                 bridge_transition: false,
                 bridge_deck_level: 0,
                 bridge_layer: None,
-                bridge_facts: crate::map::bridge_facts::BridgeCellFacts::default(),
+                bridge_facts: crate::map::bridge_facts::BridgeCellFacts {
+                    raw_flags: if ry == 2 && (1..=3).contains(&rx) {
+                        crate::map::bridge_facts::BRIDGE_FLAG_ANCHOR_SELF
+                    } else {
+                        0
+                    },
+                    ..Default::default()
+                },
                 tube_index: None,
                 radar_left: [0, 0, 0],
                 radar_right: [0, 0, 0],
