+++ src/sim/trigger_runtime_tests.rs
@@ -613,9 +613,9 @@ fn master_frame_polls_triggers_before_logic_houses_commit_and_delete() {
         vec![
             MasterFrameTestRung::SessionCommands,
             MasterFrameTestRung::Triggers,
+            MasterFrameTestRung::TeamScript,
             MasterFrameTestRung::LogicVector,
             MasterFrameTestRung::Houses,
-            MasterFrameTestRung::TeamScript,
             MasterFrameTestRung::FrameCommit,
             MasterFrameTestRung::PendingDelete,
         ]
